Move protocol version specific code into separate files
